update setting section

+++ app/Http/Controllers/Admin/AdminAuthorController.php
@@ -112,7 +112,10 @@ class AdminAuthorController extends Controller
                 'photo' => 'image|mimes:jpg,jpeg,png,gif'
             ]);
 
-            unlink(public_path('uploads/' . $author_data->photo));
+            if ($author_data->photo != NULL) {
+                unlink(public_path('uploads/' . $author_data->photo));
+            }
+
 
             $now = time();
             $ext = $request->file('photo')->extension();
